[cleanup] Move Codelite/XML/* into WebTools/XML/
